+++ src/main/kotlin/com/assari/voicebooklm/infrastructure/postgres_jpa/user/UserJpaRepository.kt
@@ -1,42 +1,16 @@
 package com.assari.voicebooklm.infrastructure.postgres_jpa.user
 
-import com.assari.voicebooklm.domain.model.User
-import com.assari.voicebooklm.domain.repository.UserRepository
+import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 import java.util.UUID
 
 /**
- * ユーザーリポジトリ実装
+ * Spring Data JPA ユーザーリポジトリインターフェース
  *
- * Domain Layer の UserRepository インターフェースを実装する。
- * Infrastructure Layer に属し、JPA を使用してデータベースアクセスを行う。
+ * 内部使用専用。外部からは UserRepository インターフェースを使用する。
  */
 @Repository
-class UserJpaRepository(
-    private val jpaRepo: SpringDataUserRepository
-) : UserRepository {
-
-    override fun save(user: User): User {
-        val entity = UserEntity.fromDomain(user)
-        val savedEntity = jpaRepo.save(entity)
-        return savedEntity.toDomain()
-    }
-
-    override fun findById(id: UUID): User? {
-        return jpaRepo.findById(id)
-            .map { it.toDomain() }
-            .orElse(null)
-    }
-
-    override fun findByEmail(email: String): User? {
-        return jpaRepo.findByEmail(email)?.toDomain()
-    }
-
-    override fun findByGoogleSub(googleSub: String): User? {
-        return jpaRepo.findByGoogleSub(googleSub)?.toDomain()
-    }
-
-    override fun deleteById(id: UUID) {
-        jpaRepo.deleteById(id)
-    }
+interface UserJpaRepository : JpaRepository<UserEntity, UUID> {
+    fun findByEmail(email: String): UserEntity?
+    fun findByGoogleSub(googleSub: String): UserEntity?
 }
